+++ lumin/nn/optimisers/__init__.py
@@ -1,4 +1,0 @@
-# from .radam import *  # noqa F304
-# from .ranger import *  # noqa F304
-
-# __all__ = [*radam.__all__, *ranger.__all__]  # noqa F405